common-utils: Move OS-related functions to the `Os` class
